requirements.txt
openai>=1.0.0
tenacity>=8.2.0
python-dotenv>=1.0.0
yfinance>=0.2.0
pandas>=1.3.0
fastapi>=0.100.0
email-validator>=2.0
uvicorn[standard]>=0.22.0
# LangChain：多步骤推理编排、外部数据接入、上下文管理；Pydantic 用于结构化输出
# 须 pydantic>=2：langchain-core 依赖 pydantic.v1（仅 v2 提供）；勿与 pydantic 1.x 并存
pydantic>=2.0,<3
langchain-core>=0.3.0
langchain-openai>=0.2.0
# RAG：向量库与检索增强（Ollama 回退需 requests）
chromadb>=0.4.0
requests>=2.28.0
# A股选股池（可选）：沪深300 / 全A按市值
akshare>=1.12.0
# 每日报告脚本（可选）：周末/节假日跳过，pip install chinesecalendar
chinesecalendar>=1.9.0
pytest>=7.0.0